# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP proxy с управлением доступом по IP (whitelist/blacklist/graylist), rate limiting (RPS/RPM/RPH/RPD/bandwidth/соединения) и мониторингом в реальном времени.",
        "title": "Proxy API",
        "contact": {
            "name": "API Support",
            "email": "artiomk00o@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/ip/blacklist": {
            "get": {
                "description": "Возвращает все записи из выбранного списка (IP-адреса, CIDR-подсети, диапазоны).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Получить IP-список",
                "responses": {
                    "200": {
                        "description": "Карта записей: ключ — IP/CIDR/диапазон, значение — true",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет IP-адрес, CIDR-подсеть (например 10.0.0.0/8) или диапазон (192.168.1.1-192.168.1.254) в список. Дубликаты игнорируются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Добавить запись в IP-список",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись добавлена, false — уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP/CIDR/диапазон или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет IP-адрес, CIDR-подсеть или диапазон из списка. Возвращает false, если записи не было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Удалить запись из IP-списка",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись удалена, false — не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ip/graylist": {
            "get": {
                "description": "Возвращает все записи из выбранного списка (IP-адреса, CIDR-подсети, диапазоны).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Получить IP-список",
                "responses": {
                    "200": {
                        "description": "Карта записей: ключ — IP/CIDR/диапазон, значение — true",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет IP-адрес, CIDR-подсеть (например 10.0.0.0/8) или диапазон (192.168.1.1-192.168.1.254) в список. Дубликаты игнорируются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Добавить запись в IP-список",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись добавлена, false — уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP/CIDR/диапазон или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет IP-адрес, CIDR-подсеть или диапазон из списка. Возвращает false, если записи не было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Удалить запись из IP-списка",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись удалена, false — не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ip/whitelist": {
            "get": {
                "description": "Возвращает все записи из выбранного списка (IP-адреса, CIDR-подсети, диапазоны).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Получить IP-список",
                "responses": {
                    "200": {
                        "description": "Карта записей: ключ — IP/CIDR/диапазон, значение — true",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет IP-адрес, CIDR-подсеть (например 10.0.0.0/8) или диапазон (192.168.1.1-192.168.1.254) в список. Дубликаты игнорируются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Добавить запись в IP-список",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись добавлена, false — уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP/CIDR/диапазон или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет IP-адрес, CIDR-подсеть или диапазон из списка. Возвращает false, если записи не было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Удалить запись из IP-списка",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись удалена, false — не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}